Refactor Nurse Campaign and Health Check components; remove unused files and update sidebar links
