--- src/pages/Nurse/Campaign/HistoryCampaign/HistoryCampaign.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const HistoryCampaign = () => {
-  return <div>HistoryCampaign</div>;
-};
-
-export default HistoryCampaign;
