--- src/pages/Nurse/Campaign/RecordForm/RecordForm.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const RecordForm = () => {
-  return <div>RecordForm</div>;
-};
-
-export default RecordForm;
