+++ src/pages/Nurse/HealthCheck/HealthCheckDetail/HealthCheckDetail.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const HealthCheckDetail = () => {
+  return <div>HealthCheckDetail</div>;
+};
+
+export default HealthCheckDetail;
