+++ src/pages/Nurse/HealthCheck/HealthCheckList/HealthCheckList.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const HealthCheckList = () => {
+  return <div>HealthCheckList</div>;
+};
+
+export default HealthCheckList;
